apps/python-api/apps/python-api/requirements.txt
# FastAPI and Web Framework
fastapi==0.117.1
uvicorn[standard]==0.37.0
gunicorn==23.0.0

# Database
sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.15.0
databases==0.9.0

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
python-multipart==0.0.19

# Validation & Settings
pydantic==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1

# Rate Limiting & Caching
slowapi==0.1.9
redis==6.4.0

# HTTP Client
aiohttp==3.12.15
httpx==0.28.1

# Testing
pytest==8.3.6
pytest-asyncio==0.25.0
pytest-cov==6.1.0
pytest-mock==3.20.0
faker==26.0.0

# Development Tools
black==24.12.0
isort==5.13.2
flake8==8.0.0
mypy==1.14.1
pre-commit==4.0.1

# Monitoring & Logging
structlog==25.0.0
sentry-sdk[fastapi]==2.23.0
prometheus-client==0.22.0

# Production
psutil==6.3.0
email-validator==2.2.0

# ML & Data Processing
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.1

# Environment & Config
python-decouple==3.8
